+++ SurvivalGame/Assets/MouseInventoryUISlot.cs
@@ -4,8 +4,11 @@ using UnityEngine.InputSystem;
 public class MouseInventoryUISlot : MonoBehaviour
 {
     public Vector2 offset = new Vector2(30f, -30f);
-    public GameObject slot;
-    public InventoryUISlot referenceUISlot;
+    public GameObject inventorySlotGO;
+    public InventorySlot originInventorySlot;
+
+    private InventorySlot inventorySlot;
+    private InventoryUISlot inventoryUISlot;
 
     private CanvasGroup canvas;
 
@@ -14,11 +17,17 @@ public class MouseInventoryUISlot : MonoBehaviour
         canvas = GetComponent<CanvasGroup>();
     }
 
+    private void Start()
+    {
+        inventorySlot = new InventorySlot(null);
+        inventoryUISlot = inventorySlotGO.GetComponent<InventoryUISlot>();
+    }
+
     void Update()
     {
         if (Mouse.current != null)
         {
-            if (slot.GetComponent<InventoryUISlot>().IsDisplayingAnItemIcon())
+            if (!inventorySlot.IsEmpty())
             {
                 canvas.alpha = 1;
                 Vector2 mousePosition = Mouse.current.position.ReadValue();
@@ -31,18 +40,38 @@ public class MouseInventoryUISlot : MonoBehaviour
         }
     }
 
-    public void Clear(bool interruptedClear = false)
+    public InventoryUISlot GetInventoryUISlot()
     {
-        if (interruptedClear)
-        {
-            Debug.LogWarning("IMPLEMENT: add back to referenceUiSlot and then clear");
-        }
-        slot.GetComponent<InventoryUISlot>().ClearSlot(destroyChild: true);
-        referenceUISlot = null;
+        return inventoryUISlot;
     }
 
-    public void SetOriginReferenceSlot(InventoryUISlot inventoryUISlot)
+    public InventorySlot GetInventorySlot()
     {
-        referenceUISlot = inventoryUISlot;
+        return inventorySlot;
+    }
+
+    public GameObject GetInventorySlotGO()
+    {
+        return inventorySlotGO;
+    }
+
+    public void Clear(bool interruptedClear = false)
+    {
+        if (interruptedClear) 
+        {
+            if (!inventorySlot.IsEmpty()) // Closed while still carrying items
+            {
+                Debug.Log("Mouse right inventory force closed. restoring items to origin");
+                originInventorySlot.AddItem(inventorySlot.itemInstance);
+            }
+        }
+        inventorySlot.ClearSlot();
+        inventoryUISlot.ClearSlot(destroyChild: true);
+        originInventorySlot = null;
+    }
+
+    public void SetOriginReferenceSlot(InventorySlot slot)
+    {
+        originInventorySlot = slot;
     }
 }